initiate CHARSET negotiation with server too
Waiting for servers to start this results in some of them sending UTF-8
before it's actually negotiated, and so Envoy puts substitute characters
in. Why the servers send UTF-8 before it's negotiated is beyond me. It
seems that if we initiate it, then it works. With some servers, at
least, we send an IAC SB CHARSET REQUEST that we don't strictly need to,
but then we just accept the server's instead. So it's perfectly fine.